[rtai] [update] urt_time is now long long
int64_t is long on 64 bit machines while its long long on 32 bit ones.
This caused confusion for printf.

+++ sys/rtai/include/urt_sys_time.h
@@ -32,7 +32,7 @@
 
 URT_DECL_BEGIN
 
-typedef int64_t urt_time;
+typedef long long urt_time;
 
 #define urt_get_time rt_get_time_ns
